SUBJECT: Screenshot and Download PR done
BODY: Used the link in the repo to obtain the required parameters using
the octokit. Personal Access Token was generated for authenticated
request, PATs are generated in github account
TASK: Need to check for bugs and added features. Need to implement for
PHP and to utilize .env
Notes: Check for bugs!!